BUGFIX: MakeBackup fixed
